REFACTOR interfaces so that 'accessors' & 'actions' are implemented separately
MAKE test classes for testing interfaces.
ADD sfneal/mock-models to composer requirements to support use of the new `InterfaceTest` trait.
BUMP CHANGELOG.md.
