+++ composer.json
@@ -24,7 +24,8 @@
     "require-dev": {
         "orchestra/testbench": ">=6.7",
         "phpunit/phpunit": ">=7.5.20",
-        "scrutinizer/ocular": "^1.8"
+        "scrutinizer/ocular": "^1.8",
+        "sfneal/mock-models": ">=0.7"
     },
     "autoload": {
         "psr-4": {
